requirements.txt
PyYAML~=6.0
Flask~=2.1.3
rpi_ws281x
adafruit-circuitpython-neopixel